requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0.0
numpy
pillow
joblib
scikit-learn
langchain
openai
google-genai
requests
itsdangerous
sarvamai
python-dotenv
sentence-transformers
cachetools
sarvamai